+++ fixtures/next/vitest/__tests__/utils/add.test.js
@@ -1,0 +1,6 @@
+import { expect, test } from 'vitest';
+import add  from '../../app/utils/add';
+
+test('Test functions that import server-only', () => {
+  expect(add(1, 2)).toBe(3)
+});