+++ src/modifiers/next/index.ts
@@ -2,6 +2,7 @@ import { withTailwindcss } from './withTailwindcss';
 import withMdx from './withMdx';
 import withWindi from './withWindi';
 import withTurbopack from './withTurbopack';
+import withVitest from './withVitest';
 
 /* browser */
 export const baseFiles = [
@@ -17,5 +18,6 @@ export const modifiers = {
   withMdx,
   withTailwindcss,
   withTurbopack,
+  withVitest,
   withWindi,
 };