+++ testUtils/read.ts
@@ -3,6 +3,10 @@ import fg from 'fast-glob';
 import path from 'node:path';
 import fs from 'node:fs/promises';
 import { EFrameworks } from '@/constants/enum';
+import { TAstValue } from 'types/index';
+import { convertTextToAst } from '@/utils/ast';
+import { setBaseProject } from '@/project';
+import { baseFiles } from '@/modifiers/next';
 
 export async function readFilesToMap(
   dir: string,
@@ -25,4 +29,19 @@ export async function readFixture(framework: EFrameworks): Promise<Map<string, s
   return await readFilesToMap(
     path.join(process.cwd(), 'fixtures', framework.toString()),
   );
+}
+
+export async function astFromFixture(): Promise<[
+  Map<string, TAstValue>,
+  Map<string, TAstValue>
+]> {
+  const fullProject = convertTextToAst(
+    await readFilesToMap(
+      path.join(process.cwd(), 'fixtures', 'next'),
+    ),
+  );
+  return [
+    fullProject,
+    setBaseProject(fullProject, baseFiles),
+  ];
 }