+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [2.3.1] - 2024-01-11
+
+### Fixed
+- vitest: test import path (app)
+- remove node `fs` from code, since it is browser only
+- expose API for vitest
+
+### Added
+- vitest: [fixture] missing add util
+
+
+
+## [2.3.1] - 2024-01-11
+
+### Added
+- support for Vitest testing framework (app)
+
+
+
 ## [2.2.1] - 2024-01-06
 
 ### Added